use 404-image instead

+++ scripts/change-theme.sh
@@ -13,18 +13,24 @@ gen_conky_conf() {
 }
 
 cp_settings() {
-  cp "$HOME/Pictures/wallpapers/$2" "$HOME/Pictures/wallpapers/wallpaper.jpg"
   cp "$dir/$1/gtkrc-2.0" "$HOME/.gtkrc-2.0"
   cp "$dir/$1/gtkrc-3.0" "$HOME/.config/gtk-3.0/settings.ini"
   cp "$dir/$1/vscode-settings.json" "$HOME/.config/Code/User/settings.json"
   cp "$dir/$1/vscode-settings.json" "$HOME/.config/Code - OSS/User/settings.json"
   cp "$dir/$1/Xresources" "$HOME/.Xresources"
 
+  if [ -f "$HOME/Pictures/wallpapers/$2" ]; then
+    cp "$HOME/Pictures/wallpapers/$2" "$HOME/Pictures/wallpapers/wallpaper.jpg"
+  else
+    cp "$HOME/Pictures/wallpapers/404-wallpaper-not-found.jpg" "$HOME/Pictures/wallpapers/wallpaper.jpg"
+  fi
+
   if [ -f "/usr/share/backgrounds/grayscaled/grayscaled-$2" ];then
     sudo sed -i "s/background =.*/background = \/usr\/share\/backgrounds\/grayscaled\/grayscaled-$2/g" "/etc/lightdm/lightdm-gtk-greeter.conf"
   else
-    sudo sed -i "s/background =.*/background = \/usr\/share\/backgrounds\/wallpapers\/$2/g" "/etc/lightdm/lightdm-gtk-greeter.conf"
+    sudo sed -i "s/background =.*/background = \/usr\/share\/backgrounds\/wallpapers\/grayscaled-404-wallpaper-not-found.jpg/g" "/etc/lightdm/lightdm-gtk-greeter.conf"
   fi
+
   # sudo sed -i "s/background=.*/background=\/usr\/share\/backgrounds\/wallpapers\/$2/g" "/etc/lightdm/slick-greeter.conf"
   # sudo sed -i "s/background-color=.*/background-color=$3/g" "/etc/lightdm/slick-greeter.conf"
